requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
idna==2.10
keyboard==0.13.5
lxml==4.6.2
mouse==0.7.1
MouseInfo==0.1.3
mss==6.0.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
PyAutoGUI==0.9.50
pygame==1.9.6
PyGetWindow==0.0.8
pyglet==1.5.7
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==228
pywinauto==0.6.8
requests==2.24.0
six==1.15.0
slacker==0.14.0
soupsieve==2.1
urllib3==1.25.10
